Moved scala-2 specific code to scala-2 source folder
