+++ core/src/main/scala-2/medeia/generic/CoproductDecoderInstances.scala
