+++ core/src/main/scala-2/medeia/generic/ShapelessEncoder.scala
